chore : change script location
