+++ pages/_app.page.tsx
@@ -1,13 +1,16 @@
 import React, { useEffect } from 'react';
 import type { AppProps } from 'next/app';
+
 import styled from '@emotion/styled';
 import { Global } from '@emotion/react';
 import { globalStyles } from 'src/style';
-import NavBar from 'src/components/navigation/NavBar/NavBar';
+
 import { useRouter } from 'next/router';
 import { RecoilRoot } from 'recoil';
+
+import NavBar from 'src/components/navigation/NavBar/NavBar';
 import Head from 'src/components/Head';
-import Script from 'next/script';
+
 import * as gtag from 'src/lib/gtag';
 
 function MyApp({ Component, pageProps }: AppProps) {
@@ -27,24 +30,6 @@ function MyApp({ Component, pageProps }: AppProps) {
 
   return (
     <>
-      <Script
-        async
-        src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_MEASUREMENT_ID}`}
-      ></Script>
-      <Script
-        id="google-analytics"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-                  window.dataLayer = window.dataLayer || [];
-                  function gtag(){dataLayer.push(arguments);}
-                  gtag('js', new Date());
-                  gtag('config', ${gtag.GA_MEASUREMENT_ID}, {
-                    page_path: window.location.pathname,
-                  });
-                `,
-        }}
-      ></Script>
       <RecoilRoot>
         <Head />
         <Global styles={globalStyles} />
