+++ pages/_document.page.tsx
@@ -1,9 +1,29 @@
 import { Html, Head, Main, NextScript } from 'next/document';
+import Script from 'next/script';
+import * as gtag from 'src/lib/gtag';
 
 export default function Document() {
   return (
     <Html>
       <Head>
+        <Script
+          async
+          src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_MEASUREMENT_ID}`}
+        ></Script>
+        <Script
+          id="google-analytics"
+          strategy="afterInteractive"
+          dangerouslySetInnerHTML={{
+            __html: `
+                  window.dataLayer = window.dataLayer || [];
+                  function gtag(){dataLayer.push(arguments);}
+                  gtag('js', new Date());
+                  gtag('config', ${gtag.GA_MEASUREMENT_ID}, {
+                    page_path: window.location.pathname,
+                  });
+                `,
+          }}
+        ></Script>
         <link
           rel="stylesheet"
           as="style"
